feat(jenkins): install the patch utility
Install the patch utility into the Jenkins container. While it's frowned
upon to have executors on the controller, I'd have to install it on
agents anyway. This is just a personal thing, guys, please don't judge
me so much, it's using patch in a script and I don't want to use a
Docker agent just because of that...