feat(builder): add an image for Jenkins builds
Add a base image that will be used for most Jenkins builds, so they
happen inside a container instead of within the Jenkins container
itself. This still doesn't make it so it doesn't run within the built-in
node, exactly, but is a step in the right direction, I would say.