refactor(builder): rename builder to local-builder
Looks more interesting than simply "builder", doesn't it? I think so. Do
you think so? Leave your comment down in the comments section! There's a
link to the package in the description below. Don't forget to subscribe,
hit the bell icon, leave your like and show it to your friends.